Moved `hosts` from Canine to Canine-2.
This involved a lot of refactoring. I moved around a lot of the
loops (for the 'allow' and 'block' commands) so that only one shell
command is ever executed, but every host specified is still added
(or removed). Also added the '--to-ip' or '-t' flag for redirecting
to a specific IP instead of localhost.