New mvim color scheme (Smyck) and mvim config file
Turns out MacVim uses .gvimrc (instead of something that makes a
lot of sense, like .mvimrc or .macvimrc), so I now finally have a
file specifically for MacVim (on top of the usual .vimrc). So far
all it does is enable the Smyck theme (added as a submodule), since
Terminal.app does not support the 256 colors required for it.