Imgur::Set includes Enumerable, sequential saving.
You can now use standard Enumerable methods on a Set. The Image class
includes a new, accessible `filename` attribute (originally set to its
id), which determines the output file's name. The '-s' flag has been
added as well, which allows for sequential filenames when saving images,
in the order they appear on the list or page.